BorderScore WIP - initial work on handling circular arenas
Circular arenas - big problem is working out where the border _is_.
My current approach uses `cv2` to find a minimum enclosing circle, but this needs adjusting so that it runs on the whole _summed_ field map, or better still, on the ratemap.
Right now, it's running on a field-by-field basis, which is no good at all.