import astropy hack
Astropy still has an issue with searching for configuration files on import, this hack should speed it up, BUT also ensure that config files are ephemeral.

Not a problem for how it is currently used, and _probably_ not a problem in the future, since we only use a few minor, standalone features in it, and don't use it all for its intended purpose.