Remove support for 1d data in 2d array
Previously, supported 1D data given as a shape (1,n) array, i.e. a 2D array, length 1, along one axis.
This is the reason for the `.squeeze` at the beginning of `accumulate_spatial` that was causing so many problems with single-spike data.
As this was a marginal usecase anyway, I have removed this support
* Tests for this approach also removed